feat(medusa): Align auth loader with new strategies
